# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Books App API",
    "description": "Backend API for Books App — books, reviews, discussions, challenges, user library, and swipe-based recommendations.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "def-14": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "title": "ApiError"
      },
...
      "def-35": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "username",
          "avatarHue",
          "level",
          "levelTitle",
          "friendsSince",
          "mutualCount"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "avatarHue": {
            "type": "integer",
            "minimum": 0,
            "maximum": 360
          },
          "level": {
            "type": "integer"
          },
          "levelTitle": {
            "type": "string"
          },
          "friendsSince": {
            "type": "string",
            "format": "date-time"
          },
          "mutualCount": {
            "type": "integer"
          }
        },
        "title": "Friend"
      },
      "def-36": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "username",
          "avatarHue",
          "level",
          "levelTitle",
          "direction",
          "sentAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "avatarHue": {
            "type": "integer",
            "minimum": 0,
            "maximum": 360
          },
          "level": {
            "type": "integer"
          },
          "levelTitle": {
            "type": "string"
          },
          "direction": {
            "type": "string",
            "enum": [
              "incoming",
              "outgoing"
            ]
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "title": "FriendRequest"
      },
      "def-37": {
        "type": "object",
        "required": [
          "data",
          "total"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/def-35"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "title": "FriendsList"
      },
      "def-38": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "incoming",
              "outgoing"
            ],
            "properties": {
              "incoming": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/def-36"
                }
              },
              "outgoing": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/def-36"
                }
              }
            }
          }
        },
        "title": "PendingRequests"
      },
      "def-39": {
        "type": "object",
        "required": [
          "userId"
        ],
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "title": "SendFriendRequestBody"
      }
    }
  },
  "paths": {
...
    "/friends": {
      "get": {
        "operationId": "getFriends",
        "summary": "Get accepted friends list",
        "tags": [
          "friends"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-37"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          }
        }
      }
    },
    "/friends/pending": {
      "get": {
        "operationId": "getPendingRequests",
        "summary": "Get incoming and outgoing pending requests",
        "tags": [
          "friends"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-38"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          }
        }
      }
    },
    "/friends/request": {
      "post": {
        "operationId": "sendFriendRequest",
        "summary": "Send a friend request",
        "tags": [
          "friends"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-39"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-36"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          }
        }
      }
    },
    "/friends/accept/{requestId}": {
      "post": {
        "operationId": "acceptFriendRequest",
        "summary": "Accept an incoming friend request",
        "tags": [
          "friends"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "requestId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-35"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          }
        }
      }
    },
    "/friends/reject/{requestId}": {
      "post": {
        "operationId": "rejectFriendRequest",
        "summary": "Reject an incoming friend request",
        "tags": [
          "friends"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "requestId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          }
        }
      }
    },
    "/friends/{friendshipId}": {
      "delete": {
        "operationId": "removeFriend",
        "summary": "Remove a friend or cancel a pending request",
        "tags": [
          "friends"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "friendshipId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-14"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
}